pyproject.toml
[project]
name = "chainlit-chatgroq-chat-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.9.6",
    "groq>=0.37.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.1",
]
